Add files via upload
this project basically has two phase customer phase and shopkeeper phase so this helps the customer in keeping track of his past purchases of which his bills are due and the shopkeeper to shares a copy of the same. This helps the customer to keep track of his past and present purchase history and also enables him to do further purchasing or pay the bill etc.